Refactor: move notification related functionality into a separate file
* Make some functions a bit more readable
